+++ docs/source/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 v0.0.5
 ------
 
-* Dropped ``__ne__` from :class:`anom.Property` as it's not natively
+* Dropped ``__ne__`` from :class:`anom.Property` as it's not natively
   supported.